Updated start script permissions check

+++ utils.php
@@ -103,7 +103,7 @@ function verify_permissions($config) {
     if (is_writable("./") == false) { // Check to se if the controller interface's root directory is writable.
         echo "<p class=\"error\">The controller interface's root directory is not writable. Please verify the permissions of the " . getcwd() . " directory.</p>";
         $valid = false;
-    } else if (is_writable("./start.sh") == false) { // Check to see if the controller interface's start script is writable.
+    } else if (file_exists("./start.sh") == true and is_writable("./start.sh") == false) { // Check to see if the controller interface's start script is writable (if it exists).
         echo "<p class=\"error\">The start.sh script in the " . getcwd() . " directory is not writable.</p>";
         $valid = false;
     }
